projects/06-crm-role/invoice-generator/app/requirements.txt
Flask
flask-cors
pandas
numpy
openpyxl
xlrd
gunicorn
fpdf2
gspread
google-auth-oauthlib
